+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-- Update for latest Evening Star Lib
+- Fog no longer only renders if you are in Survival
+  - old behavior can be restored with the new "Fog Checks Gamemode" option
+- Glowing Eyeblossoms can now be turned off
 
 Please report any issues you find to the [github](https://github.com/vesmaybevesper/paleworld)